backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pandas>=2.1.0
openpyxl>=3.1.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.75.0
langchain>=0.0.350
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
httpx>=0.24.0